% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlienindexR.R
\name{AlienindexR}
\alias{AlienindexR}
\title{AlienindexR}
\usage{
AlienindexR(blast_output_donor, blast_output_recipient)
}
\arguments{
\item{blast_output_donor}{Input file with 12 columns in tab-delimited format}

\item{blast_output_recipient}{Input file with 12 columns in tab-delimited format}
}
\value{
dataframe
}
\description{
'AlienindexR' calculates Alien Index (AI) for genes in target species from BLAST results between this species and donor/recipient species.
}
\details{
The output by BLAST or diamond BLAST will be used as input of 'AlienindexR' and should be in a tabular format with 12 columns. 

For BLAST, the output parameter should be: -outfmt '6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore'. 

For diamond BLAST, the output parameter should be: -outfmt 6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore.
}
\examples{
test_donor <- read.table(gzfile(system.file("extdata", "test_0.txt.gz", package = "AlienindexR")), header=FALSE, sep="\t")
test_recipient <- read.table(gzfile(system.file("extdata", "test_1.txt.gz", package = "AlienindexR")), header=FALSE, sep="\t")
AlienindexR(test_donor, test_recipient)
}
\references{
1. X Chen, L Yang. Manuscript in preparation.
}
\author{
Xiao Chen
}
